+++ qm/containers/capabilities_container.py
@@ -21,6 +21,7 @@ def create_capabilities_container(
         modules=[
             "qm.program._qua_config_schema",
             "qm.program._qua_config_to_pb",
+            "qm.program.ConfigBuilder",
             "qm.api.job_manager_api",
             "qm.simulate.interface",
         ],