+++ qm/type_hinting/general.py
@@ -8,6 +8,7 @@ PathLike = Union[str, bytes, pathlib.Path, os.PathLike]  # type: ignore[type-arg
 Number = Union[int, float]
 Value = Union[Number, bool]
 NumberT = TypeVar("NumberT", int, bool, float)
+"""A generic type variable that can be used to represent a pythonic `int`, `bool` or `float`."""
 
 NumpyNumber = Union[numpy.floating, numpy.integer]  # type: ignore[type-arg]
 NumpyValue = Union[NumpyNumber, numpy.bool_]